+++ app/handlers/Handler.scala
@@ -1,6 +1,7 @@
 package handlers
 
 import lib.{ProxyRequest, ResolvedToken, Route, Server}
+import play.api.libs.ws.WSClient
 import play.api.mvc.Result
 
 import scala.concurrent.{ExecutionContext, Future}
@@ -8,6 +9,7 @@ import scala.concurrent.{ExecutionContext, Future}
 trait Handler {
 
   def process(
+    wsClient: WSClient,
     server: Server,
     request: ProxyRequest,
     route: Route,
