+++ test/handlers/GenericHandlerSpec.scala
@@ -1,110 +1,17 @@
 package handlers
 
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import akka.stream.scaladsl.{Source, StreamConverters}
-import akka.util.ByteString
-import helpers.{BasePlaySpec, MockStandaloneServer}
-import lib._
+import helpers.HandlerBasePlaySpec
+import lib.{Constants, Method}
 import play.api.libs.json._
-import play.api.mvc.{Headers, Result}
 
-import scala.concurrent.duration.{FiniteDuration, MILLISECONDS}
-
-class GenericHandlerSpec extends BasePlaySpec {
+class GenericHandlerSpec extends HandlerBasePlaySpec {
 
   import scala.concurrent.ExecutionContext.Implicits.global
 
   private[this] def genericHandler = app.injector.instanceOf[GenericHandler]
 
-  private[this] case class SimulatedResponse(
-    server: Server,
-    request: ProxyRequest,
-    result: Result,
-    body: String
-  ) {
-
-    val status: Int = result.header.status
-    val contentType: Option[String] = result.body.contentType
-    val contentLength: Option[Long] = result.body.contentLength
-
-    def header(name: String): Option[String] = {
-      result.header.headers.get(name)
-    }
-
-  }
-
-  private[this] def createProxyRequest(
-    requestMethod: Method,
-    requestPath: String,
-    body: Option[ProxyRequestBody] = None,
-    queryParameters: Map[String, Seq[String]] = Map.empty,
-    headers: Map[String, Seq[String]] = Map.empty
-  ): ProxyRequest = {
-    rightOrErrors(
-      ProxyRequest.validate(
-        requestMethod = requestMethod.toString,
-        requestPath = requestPath,
-        body = body,
-        queryParameters = queryParameters,
-        headers = Headers(
-          headers.flatMap { case (k, values) =>
-            values.map { v =>
-              (k, v)
-            }
-          }.toSeq: _*
-        )
-      )
-    )
-  }
-
-  private[this] def toString(source: Source[ByteString, _]): String = {
-    implicit val system: ActorSystem = ActorSystem("toString")
-    implicit val materializer: ActorMaterializer = ActorMaterializer()
-    val is = source.runWith(StreamConverters.asInputStream(FiniteDuration(100, MILLISECONDS)))
-    scala.io.Source.fromInputStream(is, "UTF-8").mkString
-  }
-
-  private[this] def simulate(
-    method: Method,
-    path: String,
-    serverName: String = "test",
-    queryParameters: Map[String, Seq[String]] = Map.empty
-  ): SimulatedResponse = {
-    MockStandaloneServer.withTestClient { (client, port) =>
-      val server = Server(
-        name = serverName,
-        host = s"http://localhost:$port"
-      )
-
-      val proxyRequest = createProxyRequest(
-        requestMethod = method,
-        requestPath = path,
-        queryParameters = queryParameters
-      )
-
-      val result = await(
-        genericHandler.process(
-          wsClient = client,
-          server = server,
-          request = proxyRequest,
-          route = Route(method, path),
-          token = ResolvedToken(requestId = createTestId())
-        )
-      )
-
-      SimulatedResponse(
-        server = server,
-        request = proxyRequest,
-        result = result,
-        body = toString(result.body.dataStream)
-      )
-    }
-  }
-
-
   "defaults content type to application/json" in {
-    val sim = simulate(Method.Get, "/users/1")
+    val sim = simulate(genericHandler, Method.Get, "/users/1")
     sim.status must equal(200)
     sim.contentLength must equal(Some(sim.body.length))
     sim.contentType must equal(Some("application/json; charset=utf-8"))
@@ -116,7 +23,7 @@ class GenericHandlerSpec extends BasePlaySpec {
   }
 
   "propagates redirect" in {
-    val sim = simulate(Method.Get, "/redirect/example")
+    val sim = simulate(genericHandler, Method.Get, "/redirect/example")
     sim.result.header.status must equal(303)
     sim.contentLength must equal(Some(0))
     sim.header("Location") must equal(Some("http://localhost/foo"))
@@ -125,18 +32,18 @@ class GenericHandlerSpec extends BasePlaySpec {
   }
 
   "respects provided content type" in {
-    val sim = simulate(Method.Get, "/file.pdf")
+    val sim = simulate(genericHandler, Method.Get, "/file.pdf")
     sim.result.header.status must equal(200)
     sim.contentType must equal(Some("application/pdf; charset=utf-8"))
   }
 
   "propagates 404" in {
-    val sim = simulate(Method.Get, "/non-existent-path")
+    val sim = simulate(genericHandler, Method.Get, "/non-existent-path")
     sim.result.header.status must equal(404)
   }
 
   "supports jsonp" in {
-    val sim = simulate(Method.Post, "/users")
+    val sim = simulate(genericHandler, Method.Post, "/users")
     sim.status must equal(201)
     sim.contentLength must equal(Some(sim.body.length))
     sim.contentType must equal(Some("application/json; charset=utf-8"))
@@ -144,7 +51,7 @@ class GenericHandlerSpec extends BasePlaySpec {
       Json.obj("id" -> 1)
     )
 
-    val jsonp = simulate(
+    val jsonp = simulate(genericHandler, 
       Method.Post,
       "/users",
       queryParameters = Map("callback" -> Seq("foo"))
@@ -156,10 +63,10 @@ class GenericHandlerSpec extends BasePlaySpec {
   }
 
   "supports response envelope" in {
-    val sim = simulate(Method.Post, "/users")
+    val sim = simulate(genericHandler, Method.Post, "/users")
     sim.status must equal(201)
 
-    val envelope = simulate(
+    val envelope = simulate(genericHandler, 
       Method.Post,
       "/users",
       queryParameters = Map("envelope" -> Seq("response"))
